much to change

+++ database/migrations/2022_07_26_075516_buat_tabel_produk.php
@@ -20,8 +20,8 @@ class BuatTabelProduk extends Migration
             $table->string('foto_produk')->nullable();
             $table->text('konten')->nullable();
             $table->string('deskripsi')->nullable();
-            $table->integer('harga_pokok')->nullable();
-            $table->integer('harga_jual')->nullable();
+            $table->decimal('harga_pokok')->nullable();
+            $table->decimal('harga_jual')->nullable();
             $table->enum('status', ['aktif','nonaktif'])->default('aktif');
             $table->timestamps();
         });
